Методы из DAOImpl перенес в CarDAOImpl, а сам DAOImpl удалил. Класс ReflectionUtil теперь не нужен. В DAO закомментировал метод get т.к. он не используется в CarDAOImpl.
