+++ Group_Work_4_CarServlet/src/main/java/itacademy/dao/DAO.java
@@ -17,7 +17,7 @@ public interface DAO<T> {
      * @param id идентификатор строки таблицы БД
      * @return полученный из БД объект по заданному id или null, если объект не найден
      */
-    T get(Serializable id);
+    //T get(Serializable id);
 
     /**
      * Метод возвращает список всех строк таблицы в том порядке, в котором они представлены в таблице БД
